+++ mycodo/mycodo_flask/page_routes.py
@@ -59,7 +59,6 @@ from mycodo.databases.models import Sensor
 from mycodo.databases.models import Timer
 from mycodo.databases.models import User
 
-from mycodo.devices.camera import CameraStream
 from mycodo.devices.camera import camera_record
 from mycodo.utils.system_pi import add_custom_measurements
 from mycodo.utils.system_pi import csv_to_list_of_int
@@ -115,12 +114,6 @@ def page_camera():
     form_camera = forms_misc.Camera()
     camera = Camera.query.all()
 
-    # Check if a video stream is active
-    for each_camera in camera:
-        if each_camera.stream_started and not CameraStream().is_running():
-            each_camera.stream_started = False
-            db.session.commit()
-
     if request.method == 'POST':
         if not utils_general.user_has_permission('edit_settings'):
             return redirect(url_for('page_routes.page_camera'))
@@ -129,13 +122,6 @@ def page_camera():
         mod_camera = Camera.query.filter(
             Camera.id == form_camera.camera_id.data).first()
         if form_camera.capture_still.data:
-            if mod_camera.stream_started:
-                flash(gettext(
-                    u"Cannot capture still image if stream is active."))
-                return redirect('/camera')
-            if CameraStream().is_running():
-                CameraStream().terminate_controller()  # Stop camera stream
-                time.sleep(2)
             camera_record('photo', mod_camera)
         elif form_camera.start_timelapse.data:
             if mod_camera.stream_started:
@@ -172,25 +158,10 @@ def page_camera():
                 flash(gettext(
                     u"Cannot start stream if time-lapse is active."))
                 return redirect('/camera')
-            elif CameraStream().is_running():
-                flash(gettext(
-                    u"Cannot start stream. The stream is already running."))
-                return redirect('/camera')
-            elif (not (mod_camera.camera_type == 'Raspberry Pi' and
-                       mod_camera.library == 'picamera')):
-                flash(gettext(u"Streaming is only supported with the Raspberry"
-                              u" Pi camera using the picamera library."))
-                return redirect('/camera')
-            elif Camera.query.filter_by(stream_started=True).count():
-                flash(gettext(u"Cannot start stream if another stream is "
-                              u"already in progress."))
-                return redirect('/camera')
             else:
                 mod_camera.stream_started = True
                 db.session.commit()
         elif form_camera.stop_stream.data:
-            if CameraStream().is_running():
-                CameraStream().terminate_controller()
             mod_camera.stream_started = False
             db.session.commit()
         return redirect('/camera')